Rename CustomTranslateWidget to CustomTranslate
