+++ Sources/LightPortal/Addons/CustomTranslate/CustomTranslate.php
@@ -1,26 +1,26 @@
 <?php
 
 /**
- * CustomTranslateWidget.php
+ * CustomTranslate.php
  *
- * @package CustomTranslateWidget (Light Portal)
+ * @package CustomTranslate (Light Portal)
  * @link https://custom.simplemachines.org/index.php?mod=4244
  * @author Bugo <bugo@dragomano.ru>
  * @copyright 2023 Bugo
  * @license https://spdx.org/licenses/GPL-3.0-or-later.html GPL-3.0-or-later
  *
  * @category addon
- * @version 16.04.23
+ * @version 20.05.23
  */
 
-namespace Bugo\LightPortal\Addons\CustomTranslateWidget;
+namespace Bugo\LightPortal\Addons\CustomTranslate;
 
 use Bugo\LightPortal\Addons\Plugin;
 
 if (! defined('LP_NAME'))
 	die('No direct access...');
 
-class CustomTranslateWidget extends Plugin
+class CustomTranslate extends Plugin
 {
 	public string $type = 'other';
 
@@ -30,7 +30,7 @@ class CustomTranslateWidget extends Plugin
 
 	public function init()
 	{
-		if (isset($this->context['uninstalling']) || $this->request()->has('xml') || empty($this->context['lp_custom_translate_widget_plugin']['languages']))
+		if (isset($this->context['uninstalling']) || $this->request()->has('xml') || empty($this->context['lp_custom_translate_plugin']['languages']))
 			return;
 
 		if ($this->context['browser']['is_mobile'] || $this->context['browser']['possibly_robot'] || $this->context['current_action'] === 'helpadmin')
@@ -43,23 +43,23 @@ class CustomTranslateWidget extends Plugin
 
 		$this->addInlineJavaScript('new YandexTranslate({baseLang: "' . $forumLang . '"});', true);
 
-		$this->context['ctw_languages'] = array_unique(array_merge([$forumLang], explode(',', $this->context['lp_custom_translate_widget_plugin']['languages'])));
+		$this->context['ctw_languages'] = array_unique(array_merge([$forumLang], explode(',', $this->context['lp_custom_translate_plugin']['languages'])));
 
 		$this->context['ctw_lang_titles'] = array_combine($this->langCodes, $this->langTitles);
 
-		$this->setTemplate()->withLayer('custom_translate_widget');
+		$this->setTemplate()->withLayer('custom_translate');
 	}
 
 	public function addSettings(array &$config_vars)
 	{
-		$config_vars['custom_translate_widget'][] = ['callback', 'languages', [$this, 'selectLanguages']];
+		$config_vars['custom_translate'][] = ['callback', 'languages', [$this, 'selectLanguages']];
 	}
 
 	public function selectLanguages()
 	{
 		echo (new LanguageSelect)([
-			'data'  => array_combine($this->langCodes, $this->txt['lp_custom_translate_widget']['languages_set']),
-			'value' => $this->context['lp_custom_translate_widget_plugin']['languages'] ?? ''
+			'data'  => array_combine($this->langCodes, $this->txt['lp_custom_translate']['languages_set']),
+			'value' => $this->context['lp_custom_translate_plugin']['languages'] ?? ''
 		]);
 	}
