+++ Sources/LightPortal/Addons/CustomTranslate/langs/english.php
