+++ Sources/LightPortal/Addons/CustomTranslate/script.js
